+++ ChangeLog.md
@@ -1,5 +1,12 @@
 ## change log ##
 
+* 2016-07-08: 3.0.5-alpha:
+  - Re #13 "capture status"
+    - allow to change status for registered ontology.
+      Possible values: `unstable`, `testing`, `stable`, `archaic`, plus `draft`,
+      with "resource" substituted for "term" in text descriptions.
+      Ref: https://www.w3.org/2003/06/sw-vocab-status/note.html
+
 * 2016-07-07: 3.0.5-alpha:
   - style adjustments in confirmation and selection dialogs
   - <multivalueedit>: handle these keystrokes: